+++ src/components/SignUpForm/ErrorIcon/ErrorIcon.jsx
@@ -1,5 +1,5 @@
 import icon from '../../../assets/images/icons/error.svg';
-import { IconContainer } from './InputErrorIcon.styled';
+import { IconContainer } from './ErrorIcon.styled';
 
 export default function InputErrorIcon() {
   return (
