+++ src/components/SignUpForm/ErrorIcon/ErrorIcon.styled.jsx
