+++ src/components/SignUpForm/ErrorIcon/index.js
@@ -1,0 +1,1 @@
+export { default } from './ErrorIcon';
