+++ src/components/SignUpForm/EyeHideIcon/index.js
@@ -1,0 +1,1 @@
+export { default } from './EyeHideIcon';
