+++ src/components/SignUpForm/EyeOpenIcon/EyeOpenIcon.styled.jsx
