--- src/components/SignUpForm/InputErrorIcon/index.js
@@ -1,1 +1,0 @@
-export { default } from './InputErrorIcon';
