--- src/components/SignUpForm/InputSuccessIcon/index.js
@@ -1,1 +1,0 @@
-export { default } from './InputSuccessIcon';
