--- src/components/SignUpForm/PasswordOpenButtonIcon/index.js
@@ -1,1 +1,0 @@
-export { default } from './PasswordOpenButtonIcon';
