+++ src/components/SignUpForm/SuccessIcon/SuccessIcon.styled.jsx
